requirements.txt
python-telegram-bot[job-queue]==21.6
openai==2.2.0
python-dotenv==1.0.1
supabase==2.6.0
google-api-python-client==2.86.0
google-auth==2.15.0
google-auth-oauthlib==1.0.0
backports.zoneinfo==0.2.1; python_version < "3.9"
